fix: move utils to correct place
