+++ include/RE/B/BSLightingShaderProperty.h
@@ -6,6 +6,9 @@
 
 namespace RE
 {
+	class BSShader;
+	class BSLight;
+
 	class BSLightingShaderProperty : public BSShaderProperty
 	{
 	public:
@@ -15,8 +18,14 @@ namespace RE
 
 		struct Data
 		{
-			std::uint64_t unk00;  // 00
-			std::uint64_t unk08;  // 08
+			void          Clear();
+			BSRenderPass* EmplacePass(BSShader* a_shader, BSShaderProperty* a_property, BSGeometry* a_geometry,
+				uint32_t a_technique, uint8_t a_numLights = 0, BSLight* a_light0 = nullptr, BSLight* a_light1 = nullptr,
+				BSLight* a_light2 = nullptr, BSLight* a_light3 = nullptr);
+
+			// members
+			RE::BSRenderPass* head;   // 00
+			std::uint64_t     unk08;  // 08
 		};
 		static_assert(sizeof(Data) == 0x10);
